bug 1462749: move super search missing fields to django admin
