+++ webapp-django/crashstats/manage/jinja2/manage/home.html
@@ -19,10 +19,6 @@
                 <span class="fa fa-desktop fa-5x"></span>
                 Graphics Devices
             </a></li>
-            <li><a href="{{ url('manage:supersearch_fields_missing') }}">
-                <span class="fa fa-th-list fa-5x"></span>
-                Super Search Fields
-            </a></li>
         </ul>
     </div>
 </div>
