--- webapp-django/crashstats/manage/jinja2/manage/supersearch_fields_missing.html
@@ -1,54 +1,0 @@
-{% extends "manage/base.html" %}
-
-{% block page_title %}{{ super() }} - Super Search Missing Fields{% endblock %}
-
-{% block admin_title %}{{ super() }} - Super Search Missing Fields{% endblock %}
-
-{% block site_css %}
-{{ super() }}
-<style type="text/css">
-table.data-table tr td {
-    max-width: 500px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-</style>
-{% endblock %}
-
-{% block mainbody %}
-
-<div class="panel">
-    <div class="title">
-        <h2>Missing Super Search Fields</h2>
-    </div>
-    <div class="body">
-        <p>
-            This is a list of fields that are missing from the master list
-            of Super Search fields. This is computed by taking a list of all
-            fields that exist in the mappings of the last 3 weeks, then
-            comparing that list to the list of known fields.
-        </p>
-
-        <table class="data-table">
-            <thead>
-                <tr>
-                    <th>Field</th>
-                </tr>
-            </thead>
-        {% for field in missing_fields %}
-            <tr>
-                <td>{{ field }}</td>
-            </tr>
-        {% else %}
-            <tr>
-                <td>
-                    No fields missing at the moment.
-                </td>
-            </tr>
-        {% endfor %}
-        </table>
-    </div>
-</div>
-
-{% endblock %}
