+++ webapp-django/crashstats/manage/templates/admin/socorro_admin_index.html
@@ -41,6 +41,13 @@
                 <td></td>
                 <td></td>
             </tr>
+            <tr>
+                <th scope="row">
+                    <a href="{% url 'siteadmin:supersearch_fields_missing' %}">Super search fields missing</a>
+                </th>
+                <td></td>
+                <td></td>
+            </tr>
         </table>
     </div>
     {# END SECTION #}
