+++ webapp-django/crashstats/manage/tests/test_admin_views.py
@@ -1,8 +1,13 @@
+import mock
+import pytest
+
 from django.core.urlresolvers import reverse
 
 from crashstats.crashstats.tests.test_views import BaseTestViews
-
-import pytest
+from crashstats.supersearch.models import (
+    SuperSearchFields,
+    SuperSearchMissingFields,
+)
 
 
 class SiteAdminTestViews(BaseTestViews):
@@ -50,3 +55,52 @@ class TestAnalyzeModelFetches(SiteAdminTestViews):
         self._login()
         response = self.client.get(url)
         assert response.status_code == 200
+
+
+class TestSuperSearchFieldsMissing(SiteAdminTestViews):
+    def test_supersearch_fields_missing(self):
+        url = reverse('siteadmin:supersearch_fields_missing')
+        response = self.client.get(url)
+        assert response.status_code == 302
+
+        self._login()
+
+        def mocked_supersearchfields(**params):
+            return {
+                'product': {
+                    'name': 'product',
+                    'namespace': 'processed_crash',
+                    'in_database_name': 'product',
+                    'query_type': 'enum',
+                    'form_field_choices': None,
+                    'permissions_needed': [],
+                    'default_value': None,
+                    'is_exposed': True,
+                    'is_returned': True,
+                    'is_mandatory': False,
+                }
+            }
+
+        def mocked_supersearchfields_get_missing_fields(**params):
+            return {
+                'hits': [
+                    'field_a',
+                    'namespace1.field_b',
+                    'namespace2.subspace1.field_c',
+                ],
+                'total': 3
+            }
+
+        supersearchfields_mock_get = mock.Mock()
+        supersearchfields_mock_get.side_effect = mocked_supersearchfields
+        SuperSearchFields.get = supersearchfields_mock_get
+
+        SuperSearchMissingFields.implementation().get.side_effect = (
+            mocked_supersearchfields_get_missing_fields
+        )
+
+        response = self.client.get(url)
+        assert response.status_code == 200
+        assert 'field_a' in response.content
+        assert 'namespace1.field_b' in response.content
+        assert 'namespace2.subspace1.field_c' in response.content
